주식 등록, 수정, 삭제 로직을 구현하였습니다.

+++ src/main/java/com/yuha/asset_manager/entity/Asset.java
@@ -12,9 +12,10 @@ import lombok.Setter;
 public class Asset {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int stockId;
+    private long stockId;
 
     private String ticker;
+    //private String stockName;
     private int price;
     private int quantity;
